New smoother version of the GF logo
The new logo is in the new directory doc/Logos. The old logo has been moved
there too. The documentation has been updated to refer to the new logo.

The new logo was provided in SVG format by Krasimir, but PNG files are included
as well. There are some new rules in doc/Makefile to convert from SVG to PNG,
which you can use if you edit the SVG files.