(gf2ud) allow multiple labels files
sometimes one may wish to extend the existing RGL labels files using
one's own labels file. the naive way of doing this would be
copy-pasting the RGL labels file into a new file, and using that, but
then keeping it in sync would be bothersome.

the implementation in this commit is very simple: it makes the GF
shell accept (and use) several -abslabels and -cnclabels, and the file
contents are concatenated before being parsed.

warning1: I've tested only supplying an additional -abslabels file,
and it seemed to work. this is a quick hack, I can do more testing if
it is deemed necessary.

warning2: I'm still getting used to the gf-core
codebase, so if there is a better way of doing things, please let me
know.